consolidate sdss code
